Description and file delete tweaks.

+++ app/Paperroll/Helper/File.php
@@ -18,6 +18,7 @@ class File {
     }
 
     public static function delTree($dir) {
+    	if(is_link($dir)) return true;
         $files = array_diff(scandir($dir), array('.','..'));
         foreach ($files as $file) {
             (is_dir("$dir/$file")) ? self::delTree("$dir/$file") : unlink("$dir/$file");
